refactor(handler): split handlers.go into domain-specific files
Split the monolithic handlers.go into separate files per domain:
comments.go, posts.go, profiles.go, users.go, notifications.go, reactions.go.
handlers.go now only contains the Handler struct, New(), and shared response helpers.
